+++ lib/shelf_session.dart
@@ -1,4 +1,4 @@
 export 'middleware/cookies_middleware.dart';
 export 'middleware/session_middleware.dart';
 export 'session/session.dart';
-export 'session/session_storage.dart';
+export 'storage/session_storage.dart';
