+++ lib/storage/memory_storage.dart
@@ -1,0 +1,25 @@
+import 'package:shelf_session/storage/session_storage.dart';
+
+import '../session/session.dart';
+
+class MemoryStorage implements SessionStorage {
+  final Map<String, Session> _sessions = {};
+
+  @override
+  void saveSession(Session session, String sessionId) => _sessions[sessionId] = session;
+
+  @override
+  void deleteSession(String sessionId) => _sessions.remove(sessionId);
+
+  @override
+  Session? getSession(String sessionId) => _sessions[sessionId];
+
+  @override
+  bool sessionExist(String sessionId) => _sessions.containsKey(sessionId);
+
+  @override
+  void clearOutdated() {
+    final now = DateTime.now();
+    _sessions.removeWhere((k, v) => v.expires.compareTo(now) < 0);
+  }
+}
