+++ lib/storage/plain_text_storage.dart
@@ -4,41 +4,8 @@ import 'dart:io';
 
 import 'package:path/path.dart' as path;
 
-import 'session.dart';
-
-abstract class SessionStorage {
-  FutureOr<void> saveSession(Session session, String sessionId);
-
-  FutureOr<void> deleteSession(String sessionId);
-
-  FutureOr<Session?> getSession(String sessionId);
-
-  FutureOr<bool> sessionExist(String sessionId);
-
-  FutureOr<void> clearOutdated();
-}
-
-class MemoryStorage implements SessionStorage {
-  final Map<String, Session> _sessions = {};
-
-  @override
-  void saveSession(Session session, String sessionId) => _sessions[sessionId] = session;
-
-  @override
-  void deleteSession(String sessionId) => _sessions.remove(sessionId);
-
-  @override
-  Session? getSession(String sessionId) => _sessions[sessionId];
-
-  @override
-  bool sessionExist(String sessionId) => _sessions.containsKey(sessionId);
-
-  @override
-  void clearOutdated() {
-    final now = DateTime.now();
-    _sessions.removeWhere((k, v) => v.expires.compareTo(now) < 0);
-  }
-}
+import '../session/session.dart';
+import 'session_storage.dart';
 
 class PlainTextStorage implements SessionStorage {
   PlainTextStorage(this.dir);
